Create train_stinet2.py
Train the STI-net with only the orientations of the acquired dataset in Graz. The model is the STI.net, with the attention block in the lattent space, and with the orientations as an input of the attention block. Additionally, the Model was optimized a little bit to be more clear to the reader.